bmic pm_runtime get: fix double-put in case of error
if pm_runtime_resume_and_get() fails it will put the device
immediately, so we shouldn't call pm_runtime_put() on failure.

We could check the return code and only put on success, but using
pm_runtime_get_sync requires a put even on failure, so it is more
appropriate for us as we have no way of dealing with failures at this
point anyway: keep ignoring return code, and only avoid double-put.

Change-Id: If37df65b40ba648e76a0a5b3a286e31694cc163a
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>