PCI: imx6: treat link down as fatal error
This is basically a revert of commit 508919d0a95c ("PCI: imx6: Mark the
link down as non-fatal error").

If we do not consider PCI link failures as error then devices which have
no PCI card plugged in will keep the PCI subsystem up to wait for later
link up, which needlessly increase power consumption on Armadillo X1 by
about 350mW.

We do not support PCI hotplug, so the previous behavior of giving up if
no link at initialization is more appropriate for us.
A better fix would be to disable PCI in dtb and only enable it when we
expect to use it (through dtbo or otherwise), but for now reverting to
the old behavior is good enough.

Measured idle current consumption at 5V:
- before this patch 458mA
- with this patch 381mA

Change-Id: I3435fea5df60d8833589e553d93369a3e2a2a8a2
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>