thermal: imx: change calculate the temperature in i.mx7
Based on a patch in the linux-4.9-at.

The reason for the correction in linux-4.9-at is to reduce the
difference with the SoC surface temperature.

This is not the correct fix, but I value compatibility with
linux-4.9-at.

This change will result in lower temperatures being obtained. This
reduces thermal shutdowns even in high temperature environments.

Link: https://github.com/atmark-techno/linux-4.9-at/commit/6b9580b2ca98dcf0d2b858849c6bab36c534d86c
Change-Id: I0d89c80a54a587968f4449dde2431cfbbb79e29a
Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>